Removed hardcoded token, added other pages and router
Removed hard coded token, added several pages as placeholders, added router
