+++ client/package.json
@@ -27,7 +27,7 @@
     "react": "17.0.1",
     "react-dom": "17.0.1",
     "react-redux": "7.2.2",
-    "react-router-dom": "5.2.0",
+    "react-router-dom": "^5.2.0",
     "redux": "4.0.5",
     "redux-thunk": "2.3.0",
     "reselect": "4.0.0",
