+++ client/src/Pages/About.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const About = () => {
+  return <div>About placeholder</div>;
+};
+
+export default About;
