+++ client/src/Pages/Contact.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Contact = () => {
+  return <div>Contact contact</div>;
+};
+
+export default Contact;
