+++ client/src/Utils/geoLocator.js
@@ -1,11 +1,3 @@
-// export default async function getLocation() {
-//   return new Promise((resolve, reject) => {
-//     navigator.geolocation.getCurrentPosition(resolve, reject);
-//   });
-// }
-
-//original location function
-
 export default function getLocation(onSuccess) {
   const successCallback = position => {
     console.log(position);
